Application now compiles as x64 with cmake. Removed project files for Visual Studio. Added macros to indicate target arch. Fixed a number of warnings.

+++ game/g_local.h
@@ -507,7 +507,7 @@ extern	int	meansOfDeath;
 
 extern	edict_t			*g_edicts;
 
-#define STRUCT_OFS( a, b ) (long)&(((a*)0)->b)
+#define STRUCT_OFS( a, b ) ((intptr_t)&(((a*)0)->b))
 
 #define	FOFS(x)     STRUCT_OFS( edict_t, x )
 #define	STOFS(x)    STRUCT_OFS( spawn_temp_t, x )
